+++ src/main/java/net/countercraft/movecraft/repair/Repair.java
@@ -3,8 +3,8 @@ package net.countercraft.movecraft.repair;
 import java.util.UUID;
 
 import net.countercraft.movecraft.repair.config.Config;
+import net.countercraft.movecraft.repair.tasks.RepairTask;
 import net.countercraft.movecraft.repair.types.RepairQueue;
-import net.countercraft.movecraft.repair.types.RepairTask;
 
 public class Repair {
     private UUID uuid;
