+++ src/main/java/net/countercraft/movecraft/repair/tasks/RepairTask.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.repair.types;
+package net.countercraft.movecraft.repair.tasks;
 
 import org.jetbrains.annotations.Nullable;
